Add RenderBench and rename to PgnBench
