+++ src/test/scala/format/pgn/RenderTest.scala
@@ -1,10 +1,9 @@
 package chess
 package format.pgn
 
-import cats.syntax.all.*
 import scala.language.implicitConversions
 
-import MoveOrDrop.*
+import PgnHelper.*
 
 class RenderTest extends munit.FunSuite:
 
@@ -51,62 +50,6 @@ class RenderTest extends munit.FunSuite:
   test("wcc2023 study round trip tests"):
     val results = wcc2023.map(Parser.full(_).toOption.get.toPgn.render).mkString("\n\n")
     assertEquals(results, ouput2)
-
-  private def makeGame(tags: Tags) =
-    val g = Game(
-      variantOption = tags(_.Variant) flatMap chess.variant.Variant.byName,
-      fen = tags.fen
-    )
-    g.copy(
-      startedAtPly = g.ply,
-      clock = tags.clockConfig map Clock.apply
-    )
-
-  case class Context(sit: Situation, ply: Ply)
-
-  extension (d: PgnNodeData)
-    def toMove(context: Context): Option[(Situation, Move)] =
-      d.san(context.sit)
-        .toOption
-        .map(x =>
-          (
-            x.situationAfter,
-            Move(
-              ply = context.ply,
-              san = x.toSanStr,
-              comments = d.comments,
-              glyphs = d.glyphs,
-              opening = None,
-              result = None,
-              secondsLeft = None,
-              variationComments = d.variationComments
-            )
-          )
-        )
-
-  extension (tree: ParsedPgnTree)
-    def toPgn(game: Game): Option[PgnTree] =
-      tree.mapAccumlOption_(Context(game.situation, game.ply + 1)) { (ctx, d) =>
-        d.toMove(ctx) match
-          case Some((sit, m)) => (Context(sit, ctx.ply.next), m.some)
-          case None           => (ctx, None)
-      }
-
-  extension (pgn: ParsedPgn)
-    def toPgn: Pgn =
-      val game = makeGame(pgn.tags)
-      Pgn(
-        tags = pgn.tags,
-        initial = pgn.initialPosition,
-        tree = pgn.tree.flatMap(_.toPgn(game))
-      )
-
-    def cleanTags: ParsedPgn =
-      pgn.copy(tags = Tags.empty)
-
-  extension (pgn: PgnStr)
-    def cleanup = PgnStr:
-      pgn.value.replace("\r", "").replace("\n", " ").trim
 
   val pgn1 = """
 [Result "*"]
